fix : 지출 내역이 최신순으로 정렬되도록 수정 및 database query 잘못된 부분 수정 #35
