+++ app/src/main/java/com/seom/accountbook/ui/components/history/HistoryList.kt
@@ -1,0 +1,9 @@
+package com.seom.accountbook.ui.components.history
+
+import androidx.compose.runtime.Composable
+import com.seom.accountbook.model.history.HistoryModel
+import java.time.LocalDate
+
+/**
+ * 수입/지출 일별 내역을 보여주는 리스트 component
+ */
