+++ app/src/main/java/com/seom/accountbook/ui/screen/detail/DetailScreen.kt
@@ -1,3 +1,5 @@
+@file:OptIn(ExperimentalFoundationApi::class)
+
 package com.seom.accountbook.ui.screen.detail
 
 import android.graphics.Paint
@@ -29,7 +31,10 @@ import com.seom.accountbook.R
 import com.seom.accountbook.model.graph.OutComeByMonth
 import com.seom.accountbook.model.history.HistoryModel
 import com.seom.accountbook.model.history.HistoryType
+import com.seom.accountbook.ui.components.BackButtonAppBar
 import com.seom.accountbook.ui.components.OneButtonAppBar
+import com.seom.accountbook.ui.components.common.BaseDivider
+import com.seom.accountbook.ui.components.graph.LinearGraph
 import com.seom.accountbook.ui.screen.graph.GraphUiState
 import com.seom.accountbook.ui.screen.history.HistoryListHeader
 import com.seom.accountbook.ui.theme.ColorPalette
@@ -76,20 +81,13 @@ fun DetailBody(
 ) {
     Scaffold(
         topBar = {
-            OneButtonAppBar(title = title) {
-                Image(
-                    painter = painterResource(id = R.drawable.ic_back),
-                    contentDescription = null,
-                    modifier = Modifier.clickable { onBackButtonPressed() })
-            }
+            BackButtonAppBar(
+                title = title,
+                onClickBackBtn = onBackButtonPressed
+            )
         }
     ) {
-
         Column {
-            Divider(
-                color = ColorPalette.Purple,
-                thickness = 1.dp
-            )
             LinearGraph(
                 data = outComeByMonth,
                 modifier = Modifier.fillMaxWidth(),
@@ -108,10 +106,7 @@ fun DetailBody(
                     ).toString()
                 }
             )
-            Divider(
-                color = ColorPalette.LightPurple,
-                thickness = 1.dp
-            )
+            BaseDivider(color = ColorPalette.LightPurple)
             Spacer(modifier = Modifier.height(16.dp))
             OutComeList(
                 historyGroupedByDate = accounts.groupBy { LocalDate.of(it.year, it.month, it.date) }
@@ -120,98 +115,9 @@ fun DetailBody(
     }
 }
 
-val blockNum = 6
-val linearGraphPadding = 80f
-
-@Composable
-fun LinearGraph(
-    data: List<OutComeByMonth>,
-    keys: List<String>,
-    modifier: Modifier = Modifier,
-    backgroundColor: Color, // 배경 색
-    lineColor: Brush, // 선 색
-    primaryTextColor: Color, // 강조 글자 색
-    secondaryTextColor: Color // 일반 글자 색
-) {
-    val usedData = keys.map { name -> data.find { it.name == name }?.count ?: 0  }
-    Column(
-        modifier = Modifier
-            .height(130.dp)
-            .background(backgroundColor)
-            .padding(7.dp)
-    ) {
-        if (data.isNotEmpty()) {
-            Canvas(
-                modifier
-                    .weight(1f)
-                    .padding(top = 20.dp, bottom = 20.dp)
-            ) {
-                val canvasWidth = size.width
-                val canvasHeight = size.height
-
-                val minValue = usedData.minOf { it }
-                val maxValue = usedData.maxOf { it }
-                val diff = (maxValue - minValue)
-
-                val blockWidth = canvasWidth / (blockNum * 2)
-                val pointX = (0..5).map { index -> (blockWidth) * (index * 2 + 1) }
-                val pointY = usedData.map { canvasHeight - (it - minValue) / (diff / canvasHeight) }
-
-                for (index in 1 until usedData.size) {
-                    val start = Offset(x = pointX[index], y = pointY[index])
-                    val end = Offset(x = pointX[index - 1], y = pointY[index - 1])
-                    drawLine(
-                        start = start,
-                        end = end,
-                        brush = lineColor,
-                        strokeWidth = (2.0).toFloat().dp.toPx()
-                    )
-                }
-
-                keys.forEachIndexed { index, name  ->
-                    drawIntoCanvas {
-                        val count = data.find { it.name == name }?.count ?: 0
-                        it.nativeCanvas.drawText(
-                            count.toString(),
-                            pointX[index],
-                            pointY[index] - (8.sp).toPx(),
-                            Paint().apply {
-                                textSize = (10.sp).toPx()
-                                color =
-                                    (if (index == keys.size - 1) primaryTextColor else secondaryTextColor).toArgb()
-                                typeface =
-                                    Typeface.create(Typeface.DEFAULT, Typeface.BOLD)
-                                textAlign = Paint.Align.CENTER
-                            }
-                        )
-                    }
-                }
-            }
-        }
-        Row(
-            modifier = Modifier
-                .fillMaxWidth(),
-            horizontalArrangement = Arrangement.SpaceAround,
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            keys.forEachIndexed { index, row ->
-                Text(
-                    text = row,
-                    style = MaterialTheme.typography.subtitle1.copy(
-                        fontWeight = FontWeight(700),
-                        color = if (index == keys.size - 1) primaryTextColor else secondaryTextColor
-                    )
-                )
-            }
-        }
-    }
-}
-
 @RequiresApi(Build.VERSION_CODES.O)
-@OptIn(ExperimentalFoundationApi::class)
 @Composable
 fun OutComeList(
-    // TODO Key 를 String 으로 두는게 맞을까...
     historyGroupedByDate: Map<LocalDate, List<HistoryModel>>,
     modifier: Modifier = Modifier
 ) {
